we-move-save-get: Move weight save and get_weights_by_name methods onto WeightExporter
Refactor chain ID: we-move-save-get
