+++ python/sglang/srt/managers/tp_worker.py
@@ -180,7 +180,7 @@ class BaseTpWorker(ABC):
         return success, message
 
     def get_weights_by_name(self, recv_req: GetWeightsByNameReqInput):
-        parameter = self.model_runner.get_weights_by_name(
+        parameter = self.model_runner.weight_exporter.get_weights_by_name(
             recv_req.name, recv_req.truncate_size
         )
         return parameter
